Remove custom errors and replace with a Message class
Errors are to be handled with the appropriate error message shown
to the client instead of showing stack traces. Application features
and tests have been updated to reflect the use of the new messages
module.  Testing has also been improved through removing hard-coded
messages and instead relying on the Messages module to display the
correct message, decreasing the fragility of the tests and creating a
single source of truth for messages to the client.

Other minor improvements include refactoring tests through additional
fixtures and including additional marks to improve testing.